+++ lib/src/data_preprocessing/data_frame/csv_codec_factory/csv_codec_factory.dart
@@ -1,0 +1,8 @@
+import 'package:csv/csv.dart';
+
+abstract class CsvCodecFactory {
+  CsvCodec create({
+    String fieldDelimiter,
+    String eol,
+  });
+}